+++ src/handler/userHandler.js
@@ -2,33 +2,41 @@ import User from "../models/User";
 
 export const getUserList = async (req, res) => {
   const users = await User.find({});
-  return res.json(users);
+  if (users === null) {
+    return res.status(404);
+  }
+  return res.status(200).json(users);
 };
 
 export const getUser = async (req, res) => {
   const _id = req.params.id;
   if (_id === null) {
-    return res.json({ message: "no params" });
+    return res.status(400).json({ message: "no params" });
   }
   try {
     const user = await User.findOne({ _id });
-    return res.json({ message: "ok", user });
+    if (!user) {
+      return res.status(404).json({ message: "no account" });
+    }
+    return res.status(200).json({ message: "ok", user });
   } catch (err) {
     console.log("에러");
-    return res.json({ message: { err } });
+    return res.status(500).json({ message: { err } });
   }
 };
 
 export const getDeleteUser = async (req, res) => {
   console.log("deleteUser get 요청이 옴");
   const _id = req.params.id;
-  console.log(_id);
+  if (_id === null) {
+    return res.status(400);
+  }
   try {
     await User.deleteOne({ _id });
     console.log("삭제완료");
-    return res.json({ message: "ok" });
+    return res.status(204).json({ message: "ok" });
   } catch (err) {
     console.log("에러뜸");
-    return res.json({ message: { err } });
+    return res.status(500).json({ message: { err } });
   }
 };
